@@ -3,14 +3,11 @@
   "version": "0.1.1",
   "private": true,
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "6.1.1",
-    "@fortawesome/free-solid-svg-icons": "6.1.1",
-    "@fortawesome/react-fontawesome": "0.1.18",
     "axios": "0.27.2",
-    "bootstrap": "5.1.3",
-    "react": "18.1.0",
-    "react-dom": "18.1.0",
-    "react-router-dom": "6.3.0"
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-router-dom": "6.3.0",
+    "web-vitals": "2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -19,7 +16,10 @@
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
-    "extends": "react-app"
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
   },
   "browserslist": {
     "production": [
@@ -35,14 +35,13 @@
   },
   "devDependencies": {
     "@testing-library/jest-dom": "5.16.4",
-    "@testing-library/react": "13.2.0",
-    "@testing-library/user-event": "14.2.0",
-    "@types/jest": "27.5.1",
-    "@types/node": "16.11.35",
-    "@types/react": "18.0.9",
-    "@types/react-dom": "18.0.4",
-    "@types/react-router-dom": "5.3.3",
+    "@testing-library/react": "13.3.0",
+    "@testing-library/user-event": "14.2.1",
+    "@types/jest": "27.5.2",
+    "@types/node": "16.11.43",
+    "@types/react": "18.0.15",
+    "@types/react-dom": "18.0.6",
     "react-scripts": "5.0.1",
-    "typescript": "4.6.4"
+    "typescript": "4.7.4"
   }
 }
